Added description on how to obtain dictionary.txt
http://www-01.sil.org/linguistics/wordlists/english/wordlist/wordsEn.txt
Download this file and put it in the same directory as Mars and the main .asm file. Rename it dictionary.txt as well. The program should work after that.
It will: Read in a random word from the wordlist, then the player will have to guess what letters are in the word. When three wrong guesses are made, the game ends. When all the letters are guessed, the game ends in success for the player. Program also contains Tyler's sounds posted yesterday.